refactor(core): expose library entrypoint
